Moved most of the bluetooth code in esp32 to it's own file
